Created test_similarity_predictor.py unit test Return SimilarityDebugInfo() in column similarity if table similarity score is less than threshold updated  table_similarity_labels_IDEAL.json created demo_similarity_check.py
