+++ datahub-classify/tests/expected_output/table_similarity_labels_IDEAL.json
@@ -3079,5 +3079,37 @@
     "vehicles_5-vehicles_5_LOGICAL_COPY": "similar",
     "VISA-1664534613-VISA-1664534613_LOGICAL_COPY": "similar",
     "visitor-interests-visitor-interests_LOGICAL_COPY": "similar",
-    "Voyager-1664535043-Voyager-1664535043_LOGICAL_COPY": "similar"
+    "Voyager-1664535043-Voyager-1664535043_LOGICAL_COPY": "similar",
+    "1-MB-Test-1-MB-Test_COPY": "similar",
+    "2018-seattle-business-districts-2018-seattle-business-districts_COPY": "similar",
+    "Age2_address1_credit_card3-Age2_address1_credit_card3_COPY": "similar",
+    "Airbnb_Open_Data-Airbnb_Open_Data_COPY": "similar",
+    "Aliases-Aliases_COPY": "similar",
+    "application_record-application_record_COPY": "similar",
+    "Bachelor_Degree_Majors-Bachelor_Degree_Majors_COPY": "similar",
+    "Book1.xlsx-credit-card-number-Book1.xlsx-credit-card-number_COPY": "similar",
+    "Book1.xlsx-us-social-security-22-cvs-Book1.xlsx-us-social-security-22-cvs_COPY": "similar",
+    "CardBase-CardBase_COPY": "similar",
+    "catalog-catalog_COPY": "similar",
+    "CrabAgePrediction-CrabAgePrediction_COPY": "similar",
+    "Credit_Card2-Credit_Card2_COPY": "similar",
+    "Customer_Segmentation-Customer_Segmentation_COPY": "similar",
+    "drug-use-by-age-drug-use-by-age_COPY": "similar",
+    "Emails-Emails_COPY": "similar",
+    "iban-iban_COPY": "similar",
+    "ipv6_random_generated-ipv6_random_generated_COPY": "similar",
+    "list_of_real_usa_addresses-list_of_real_usa_addresses_COPY": "similar",
+    "Persons-Persons_COPY": "similar",
+    "random_ibans-random_ibans_COPY": "similar",
+    "Salary_Data-Salary_Data_COPY": "similar",
+    "sample-data-sample-data_COPY": "similar",
+    "test-test_COPY": "similar",
+    "train-train_COPY": "similar",
+    "UCI_Credit_Card-UCI_Credit_Card_COPY": "similar",
+    "US_Driving_License-US_Driving_License_COPY": "similar",
+    "vehicles_1-vehicles_1_COPY": "similar",
+    "vehicles_2-vehicles_2_COPY": "similar",
+    "vehicles_3-vehicles_3_COPY": "similar",
+    "vehicles_4-vehicles_4_COPY": "similar",
+    "vehicles_5-vehicles_5_COPY": "similar"
 }
